compile out ndk for unsupported platforms
