+++ InworldAI/Source/InworldAIClient/InworldAIClient.Build.cs
@@ -39,7 +39,6 @@ public class InworldAIClient : ModuleRules
             PrivateDependencyModuleNames.AddRange(
                 new string[]
                 {
-                    "InworldAINDK",
                     "InworldAINDKLibrary",
                 });
         }
