+++ InworldAI/Source/InworldAIClient/Private/InworldPacketTranslator.cpp
@@ -5,6 +5,7 @@
  * that can be found in the LICENSE.md file or at https://www.inworld.ai/sdk-license
  */
 
+#ifdef INWORLD_WITH_NDK
 #include "InworldPacketTranslator.h"
 THIRD_PARTY_INCLUDES_START
 #include "Utils/Utils.h"
@@ -193,3 +194,5 @@ void InworldPacketTranslator::TranslateEvent<Inworld::RelationEvent, FInworldRel
 	New.Respect = Original.GetRespect();
 	New.Trust = Original.GetTrust();
 }
+
+#endif
