+++ InworldAI/Source/InworldAIClient/Private/InworldPacketTranslator.h
@@ -9,6 +9,7 @@
 
 #include "InworldPackets.h"
 
+#ifdef INWORLD_WITH_NDK
 THIRD_PARTY_INCLUDES_START
 #include "Packets.h"
 THIRD_PARTY_INCLUDES_END
@@ -53,3 +54,5 @@ protected:
 		Packet = NewPacket;
 	}
 };
+
+#endif
